Fix: fake data and temporary find schedules

+++ src/routes/entities/route.entity.ts
@@ -16,7 +16,7 @@ export class Route {
   @Column({ type: 'integer', nullable: true })
   distanceKm: number
 
-  @Column({ type: 'integer', nullable: true })
+  @Column({ type: 'float', nullable: true })
   durationHours: number
 
   @OneToMany(() => RouteStop, (routeStop) => routeStop.route)
